Allow custom sync config locations
Add a check for the environment variable "SYNC_CONFIG" which is
used to find a configuration file. This allows to place the files
at a different location than the acutal project, making re-use of
configurations easier.

Example

SYNC_CONFIG="$HOME/sync-projecta.yml" ct sync:server staging

The default behaviour is not altered by this change.